+++ src/util/actor.test.ts
@@ -186,4 +186,18 @@ describe('Actor', () => {
 
         expect(spy).not.toHaveBeenCalled();
     });
+
+    test('should process a message when origin is "file://"', async () => {
+        const worker = workerFactory() as any as WorkerGlobalScopeInterface & ActorTarget;
+        const actor = new Actor(worker, '1');
+
+        const spy = jest.fn().mockReturnValue(Promise.resolve({}));
+        worker.worker.actor.registerMessageHandler('getClusterExpansionZoom', spy);
+
+        actor.target.postMessage({type: 'getClusterExpansionZoom', data: {} as any, origin: 'file://'});
+
+        await sleep(0);
+
+        expect(spy).toHaveBeenCalled();
+    });
 });